refactor(skintillator): use functional approach to load
