+++ src/scintillator/index.js
@@ -5,40 +5,31 @@ let debug = Debug('scintillator:loader')
 import $ from 'jquery'
 import url from 'url'
 import promiseForPixi from './import-pixi'
+import co from 'co'
 
 import Resources from './resources'
 import Compiler from './compiler'
 
-/**
- * A Scintillator is Bemuse's graphics display engine based on PIXI.js.
- */
-export class Scintillator {
-  constructor({ PIXI, $xml, resources }) {
-    this._PIXI = PIXI
-    this._$xml = $xml
-    this._resources = resources
-  }
-  compile(options={}) {
-    let $xml = this._$xml
-    new Compiler({ $xml }).compile()
-    void options
-  }
-}
-
 export function load(xmlPath) {
-  debug('load XML from %s', xmlPath)
-  let promiseForXml = loadXml(xmlPath)
-  return Promise.all([promiseForPixi, promiseForXml])
-  .then(([PIXI, $xml]) => {
-    debug('PIXI and $xml loaded')
+  return co(function*() {
+
+    debug('load XML from %s', xmlPath)
+    let [$xml, PIXI] = yield Promise.all([loadXml(xmlPath), promiseForPixi])
+
+    debug('loading resources')
     let resources = new Resources()
     for (let image of Array.from($xml.children('image'))) {
       let src = $(image).attr('src')
       let imageUrl = url.resolve(xmlPath, src)
       resources.add(src, imageUrl)
     }
-    return loadResources(resources).with(PIXI)
-    .then(() => new Scintillator({ PIXI, $xml, resources }))
+    yield loadResources(resources).with(PIXI)
+
+    debug('compiling')
+    let skin = new Compiler($xml).compile()
+
+    return skin
+
   })
 }
 
@@ -63,4 +54,3 @@ function loadResources(resources) {
   }
 }
 
-export default Scintillator
